code for chapter 4
