--- README.md
@@ -1,51 +1,0 @@
-# FastAPI Beyond The CRUD Stuff
-
-This is source code for a video tutorial series we will cover on FastAPI. THis will focus on fastapi API Development beyond the CRUD stuff.
-
-A comprehensive guide is available [here](./docs/index.md).
-
-## Table of Contents
-- [Introduction](#introduction)
-- [Getting Started](#getting-started)
-  - [Prerequisites](#prerequisites)
-  - [Installation](#installation)
-  - [Running the Application](#running-the-application)
-
-## Introduction
-
-Explain what your FastAPI project does and why it exists. Provide a high-level overview of its features and functionality.
-
-
-## Getting Started
-
-Provide instructions on how to set up and run your FastAPI project.
-
-### Prerequisites
-
-This project requires that you have the following dependencies
-
-- Python 3.x
-- FastAPI
-
-
-### Project setup
-1. Beging by cloning the project
-```bash
-git clone https://github.com/jod35/fastapi-beyond-the-CRUD-stuff.git
-```
-
-2. Enter your project
-```bash
-cd fastapi-beyond-the-CRUD-stuff
-```
-3. Install your FastAPI project dependencies, you will need to follow the steps below
-
-```bash
-pip install -r requirements.txt
-```
-
-#4. Finally run your the application.
-
-```bash
-uvicorn main:app --reload
-```
