+++ run.py
@@ -1,0 +1,17 @@
+import uvicorn
+
+
+
+if __name__ == '__main__':
+    uvicorn.run(
+        app="src:app",
+        reload=True,
+        port=8000
+    )
+
+
+
+
+
+
+
