+++ src/book_data.py
@@ -1,8 +1,3 @@
-from fastapi import FastAPI, Query
-from schemas import BookSchema,BookUpdateSchema
-
-app = FastAPI()
-
 books = [
     {
         "id": 1,
@@ -59,49 +54,3 @@ books = [
         "language": "English",
     },
 ]
-
-
-
-
-@app.get("/books")
-async def read_books():
-    """Read all books"""
-    return books
-
-@app.get('/book/{book_id}')
-async def read_book(book_id: int):
-    """Read a book"""
-    for book in books:
-        if book['id'] == book_id:
-            return book
-    return {"message": "Book not found"}
-
-@app.post('/books',status_code=201)
-async def create_book(book: BookSchema):
-    """Create a new book"""
-    books.append(book)
-    return book
-
-@app.patch('/book/{book_id}')
-async def update_book(book_id: int, update_data: BookUpdateSchema):
-    """"update book """
-    for book in books:
-        if book['id'] == book_id:
-            book['title'] = update_data.title
-            book['author'] = update_data.author
-            book['publisher'] = update_data.publisher
-            book['page_count'] = update_data.page_count
-            book['language'] = update_data.language
-            return book
-    return {"message": "Book not found"}
-
-
-@app.delete('/book/{book_id}',status_code=204)
-async def delete_book(book_id: int):
-    """delete a book"""
-    for book in books:
-        if book['id'] == book_id:
-            books.remove(book)
-            return book
-    return {"message": "Book not found"}
-
